+++ src/data/emacs-mode/agda2-mode.el
@@ -203,22 +203,12 @@ to this variable to take effect."
                                                     (string-to-syntax " ")))))
                         (modify-syntax-entry keys "w" tbl)))
                     (standard-syntax-table))
-    ;; Then override the remaining special cases.
-    (dolist (cs '((?{ . "(}1n") (?} . "){4n") (?- . "w 123b") (?\n . "> b")
-                  (?. . ".") (?\; . ".") (?_ . ".") (?! . ".")))
-      (modify-syntax-entry (car cs) (cdr cs) tbl))
     tbl)
-  "Syntax table used by the Agda mode:
+  "Syntax table used by the Agda mode.
 
-{}   | Comment characters, matching parentheses.
--    | Comment character, word constituent.
-\n   | Comment ender.
-.;_! | Punctuation.
-
-Remaining characters inherit their syntax classes from the
-standard syntax table if that table treats them as matching
-parentheses or whitespace.  Otherwise they are treated as word
-constituents.")
+A character inherits its syntax class from the standard syntax
+table if that table treats it as a matching parenthesis or
+whitespace. Otherwise it is treated as a word constituent.")
 
 (defconst agda2-command-table
   `(
@@ -1009,14 +999,6 @@ The buffer is returned.")
 (agda2-warning-or-info-buffer
  agda2-info-buffer "info" "*Agda information*")
 
-(defun agda2-font-syntactic-face (state)
-  (cond ((nth 4 state)
-         ( save-excursion
-           (goto-char (nth 8 state))
-           (cond ((looking-at "--[[:space:]\n]") 'font-lock-comment-face)
-                 ((looking-at "{-[^#]") 'font-lock-comment-face)
-          )))))
-
 (defun agda2-info-action (name text &optional append)
   "Insert TEXT into the Agda info buffer and display it.
 NAME is displayed in the buffer's mode line.
@@ -1536,7 +1518,7 @@ Only if the buffer is unmodified, and only if there is anything to load."
 
 (defun agda2-literate-p ()
   "Is the current buffer a literate Agda buffer?"
-  (equal (file-name-extension (buffer-name)) "lagda"))
+  (not (equal (file-name-extension (buffer-file-name)) "agda")))
 
 (defmacro agda2--case (exp &rest branches) ;FIXME: Use `pcase' instead!
   (declare (debug t) (indent 1))
@@ -1830,17 +1812,6 @@ a file is loaded."
 
 (defun agda2-comments-and-paragraphs-setup nil
   "Set up comment and paragraph handling for Agda mode."
-
-  ;; Syntax table setup for comments is done elsewhere.
-
-  ;; Enable highlighting of comments via Font Lock mode (which uses
-  ;; the syntax table).
-  (set (make-local-variable 'font-lock-defaults)
-       '(nil nil nil nil nil (font-lock-syntactic-face-function . agda2-font-syntactic-face)))
-  ;; If the following s-expression is removed, then highlighting of
-  ;; comments stops working.
-  (when font-lock-mode
-    (font-lock-mode t))
 
   ;; Empty lines (all white space according to Emacs) delimit
   ;; paragraphs.